--- src/app/auth/sign-in/_components/sign-in-form/sign-in-form.module.scss
@@ -1,28 +1,0 @@
-@use '@/styles/index' as *;
-
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 28px;
-
-  .input_container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 12px;
-    width: 100%;
-
-    .error {
-      @include Paragraph-02;
-      color: var(--brand-main);
-      display: flex;
-      justify-content: flex-start;
-      width: 100%;
-    }
-  }
-
-  .button_container {
-    display: flex;
-    width: 100%;
-  }
-}
