--- src/app/auth/sign-in/_services/sign-in-api.type.ts
@@ -1,6 +1,0 @@
-interface SignInAPIBody {
-  username: string;
-  password: string;
-}
-
-export type { SignInAPIBody };
